Fix NumVersionVariant union type-punning for 64-bit
NumVersionVariant uses a union to convert between NumVersion struct
and uint32_t whole value. Two critical bugs were fixed:

1. Union size mismatch (64-bit only):
   - `unsigned long whole` is 8 bytes on LP64 (Linux x86_64)
   - NumVersion struct is 4 bytes (4 x uint8)
   - Upper 4 bytes of `whole` were uninitialized
   - Changed to `uint32_t whole` for consistent 4-byte size

2. Union type-punning (undefined behavior):
   - Writing to one union member and reading from another is UB
   - GetVersionNumber() wrote to `whole`, read from `parts`
   - WriteToMap() read from `whole` after writing to `parts`
   - Fixed GetVersionNumber() to directly initialize NumVersion struct
   - Fixed WriteToMap() with defensive `{0}` initialization

The type-punning bug caused version corruption in preferences file:
- Version_v2_0 written as 00000000 instead of 0x04104003
- About dialog showed "Unknown" instead of "4.1a3"
- X-Mailer header was corrupted

Changes:
- Linux/Sources/Mulberry_Prefix.h: unsigned long -> uint32_t
- Win32/Sources/Mac2Win.h: unsigned long -> uint32_t (consistency)
- CMulberryAppProtect.cp: Eliminate type-punning with direct struct init
- CPreferencesRW.cp: Add defensive initialization before union use

+++ Linux/Sources/Mulberry_Prefix.h
@@ -18,6 +18,8 @@
 #ifndef _MULBERRY_PREFIX_H
 #define _MULBERRY_PREFIX_H
 
+#include <stdint.h>
+
 //The various OS's
 #define __win32_os 95
 #define __mac_os 2
@@ -58,7 +60,7 @@ typedef struct NumVersion               NumVersion;
 union NumVersionVariant {
 																/* NumVersionVariant is a wrapper so NumVersion can be accessed as a 32-bit value */
 	NumVersion 						parts;
-	unsigned long 					whole;
+	uint32_t 						whole;
 };
 
 class JRect;
